+++ CHANGELOG.md
@@ -10,11 +10,12 @@
     * Code analysis improvements and cleanup
     * Removed Twig panel from debugger (no longer supported in Twig 1.20)
     * Updated composer packages
-    * Prepend active languaeg to `convertUrl()` when used in markdown links
+    * Prepend active language to `convertUrl()` when used in markdown links
     * Added some pre/post flight options for installer via blueprints
     * Hyphenize the site name in the backup filename
 1. [](#bugfix)
     * Fix broken routable logic
+    * Check for `phpinfo()` method in case it is restricted by hosting provider
     * Fixes for windows when running GPM
     * Fix for ampersand (`&`) causing error in `truncateHtml()` via `Page.summary()`